Polishing - move and rename test (classes).
